vboot: Split partition and vblock verification from LoadKernel()
LoadKernel() was a big function which did everything from looping over
partitions on a drive to loading the data within them to calling the
low-level verification functions on that data.  Split it apart into more
manageable chunks.  This also reduces indentation of the inner parts of
the code, whic increases readability.

No outwardly-visible functionality changes.

BUG=chromium:611535
BRANCH=none
TEST=make runtests; emerge-kevin coreboot depthcharge

+++ firmware/lib/vboot_api_kernel.c
@@ -1290,9 +1290,9 @@ VbError_t VbSelectAndLoadKernel(VbCommonParams *cparams,
 
 	/* Save disk parameters */
 	kparams->disk_handle = p.disk_handle;
-	kparams->partition_number = (uint32_t)p.partition_number;
+	kparams->partition_number = p.partition_number;
 	kparams->bootloader_address = p.bootloader_address;
-	kparams->bootloader_size = (uint32_t)p.bootloader_size;
+	kparams->bootloader_size = p.bootloader_size;
 	kparams->flags = p.flags;
 	memcpy(kparams->partition_guid, p.partition_guid,
 	       sizeof(kparams->partition_guid));
